+++ geo_matcher/logger.py
@@ -32,7 +32,7 @@ class _ColorFormatter(logging.Formatter):
 
 def _get_log_level() -> int:
     """Set the logging level"""
-    level_name = os.getenv("GEO_MATCHER_DEBUG_LEVEL", "DEBUG").upper()
+    level_name = os.getenv("GEO_MATCHER_DEBUG_LEVEL", "INFO").upper()
     level = getattr(logging, level_name, None)
 
     if not isinstance(level, int):